# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RLEGS AM Monitoring & Bot Telegram API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  id: number;
  email: string;
  role: string;
}

export interface AccountManager {
  id: number;
  nik: string;
  nama: string;
  slug: string;
  divisi: string;
  /** @nullable */
  segmen?: string | null;
  witel: string;
  /** @nullable */
  telegramChatId?: string | null;
  telegramConnected: boolean;
  kpiActivity: number;
  createdAt: string;
}

export interface CreateAccountManagerBody {
  nik: string;
  nama: string;
  divisi: string;
  /** @nullable */
  segmen?: string | null;
  witel: string;
  /** @nullable */
  telegramChatId?: string | null;
  kpiActivity?: number;
}

export interface UpdateAccountManagerBody {
  /** @nullable */
  nama?: string | null;
  /** @nullable */
  divisi?: string | null;
  /** @nullable */
  segmen?: string | null;
  /** @nullable */
  witel?: string | null;
  /** @nullable */
  telegramChatId?: string | null;
  /** @nullable */
  kpiActivity?: number | null;
}

export interface ImportUrlBody {
  /** @nullable */
  url?: string | null;
  /** @nullable */
  period?: string | null;
}

export interface ImportResult {
  success: boolean;
  rowsImported: number;
  amCount: number;
  period: string;
  message: string;
  importId: number;
}

export interface ImportRecord {
  id: number;
  type: string;
  rowsImported: number;
  period: string;
  snapshotDate?: string | null;
  createdAt: string;
  autoTelegramSent: boolean;
}

export interface PerformanceSummary {
  nik: string;
  namaAm: string;
  divisi: string;
  tahun: number;
  bulan: number;
  targetRevenue: number;
  realRevenue: number;
  achRate: number;
  achRateYtd: number;
  rankAch: number;
  statusWarna: string;
  /** @nullable */
  komponenDetail?: string | null;
}

export interface PerformanceDetail {
  nik: string;
  namaAm: string;
  divisi: string;
  summaries: PerformanceSummary[];
}

export interface FunnelLop {
  lopid: string;
  judulProyek: string;
  pelanggan: string;
  nilaiProyek: number;
  divisi: string;
  statusF: string;
  /** @nullable */
  statusProyek?: string | null;
  /** @nullable */
  kategoriKontrak?: string | null;
  /** @nullable */
  estimateBulan?: string | null;
  namaAm: string;
  reportDate: string;
}

export interface FunnelStatusGroup {
  status: string;
  count: number;
  totalNilai: number;
}

export interface FunnelAmGroup {
  namaAm: string;
  nik: string;
  divisi: string;
  totalLop: number;
  totalNilai: number;
  shortage: number;
  byStatus: FunnelStatusGroup[];
}

export interface FunnelOverview {
  totalLop: number;
  totalNilai: number;
  targetFullHo: number;
  realFullHo: number;
  shortage: number;
  /** @nullable */
  cutoffDate?: string | null;
  byStatus: FunnelStatusGroup[];
  byAm: FunnelAmGroup[];
  lops: FunnelLop[];
}

export interface FunnelDetail {
  nik: string;
  namaAm: string;
  divisi: string;
  totalLop: number;
  totalNilai: number;
  shortage: number;
  lops: FunnelLop[];
}

export interface ActivityItem {
  id: number;
  nik: string;
  fullname: string;
  divisi: string;
  activityEndDate: string;
  label: string;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  caName?: string | null;
  /** @nullable */
  lopid?: string | null;
  /** @nullable */
  picName?: string | null;
  /** @nullable */
  activityNotes?: string | null;
}

export interface ActivityLabelGroup {
  label: string;
  count: number;
}

export interface ActivityAmGroup {
  nik: string;
  fullname: string;
  divisi: string;
  activityCount: number;
  kpiTarget: number;
  kpiAchieved: boolean;
  byLabel: ActivityLabelGroup[];
}

export interface ActivityOverview {
  totalActivity: number;
  byAm: ActivityAmGroup[];
  activities: ActivityItem[];
}

export interface ActivityDetail {
  nik: string;
  fullname: string;
  divisi: string;
  activityCount: number;
  kpiTarget: number;
  activities: ActivityItem[];
}

export interface SendTelegramBody {
  /** @nullable */
  targetNiks?: string[] | null;
  period: string;
  includePerformance: boolean;
  includeFunnel: boolean;
  includeActivity: boolean;
  /** @nullable */
  customMessage?: string | null;
}

export interface TelegramSendDetail {
  nik: string;
  namaAm: string;
  status: string;
  /** @nullable */
  error?: string | null;
}

export interface TelegramSendResult {
  sent: number;
  failed: number;
  skipped: number;
  details: TelegramSendDetail[];
}

export interface TelegramLog {
  id: number;
  nik: string;
  namaAm: string;
  status: string;
  period: string;
  messageType: string;
  /** @nullable */
  error?: string | null;
  createdAt: string;
}

export interface GenerateTelegramCodeBody {
  amId: number;
}

export interface TelegramCodeResponse {
  code: string;
  expiresAt: string;
}

export interface Settings {
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  sharepointPerformanceUrl?: string | null;
  /** @nullable */
  sharepointFunnelUrl?: string | null;
  /** @nullable */
  sharepointActivityUrl?: string | null;
  autoSendOnImport: boolean;
  kpiActivityDefault: number;
}

export interface UpdateSettingsBody {
  /** @nullable */
  telegramBotToken?: string | null;
  /** @nullable */
  sharepointPerformanceUrl?: string | null;
  /** @nullable */
  sharepointFunnelUrl?: string | null;
  /** @nullable */
  sharepointActivityUrl?: string | null;
  /** @nullable */
  autoSendOnImport?: boolean | null;
  /** @nullable */
  kpiActivityDefault?: number | null;
}

export interface PublicAmProfile {
  nik: string;
  nama: string;
  slug: string;
  divisi: string;
  witel: string;
  performance?: PerformanceSummary;
  funnel?: FunnelDetail;
  activity?: ActivityDetail;
}

export type ListPerformanceParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
  /**
   * @nullable
   */
  divisi?: string | null;
};

export type GetPerformanceByNikParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
};

export type ListFunnelParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
  /**
   * @nullable
   */
  divisi?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  nama_am?: string | null;
};

export type GetFunnelByNikParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
};

export type ListActivityParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
  /**
   * @nullable
   */
  divisi?: string | null;
  /**
   * @nullable
   */
  nama_am?: string | null;
  /**
   * @nullable
   */
  label?: string | null;
};

export type GetActivityByNikParams = {
  /**
   * @nullable
   */
  year?: number | null;
  /**
   * @nullable
   */
  month?: number | null;
};

export type GetPublicAmProfileParams = {
  nik: string;
};