# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RLEGS AM Monitoring & Bot Telegram API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /performance:
    get:
      operationId: listPerformance
      tags: [performance]
      summary: Get performance data for all AMs
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: divisi
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformanceSummary"
  /performance/{nik}:
    get:
      operationId: getPerformanceByNik
      tags: [performance]
      summary: Get performance data for specific AM
      parameters:
        - name: nik
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /funnel:
    get:
      operationId: listFunnel
      tags: [funnel]
      summary: Get funnel summary for all AMs
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: divisi
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: nama_am
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelOverview"
  /funnel/{nik}:
    get:
      operationId: getFunnelByNik
      tags: [funnel]
      summary: Get funnel data for specific AM
      parameters:
        - name: nik
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelDetail"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PerformanceSummary:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        divisi:
          type: string
        tahun:
          type: integer
        bulan:
          type: integer
        targetRevenue:
          type: number
        realRevenue:
          type: number
        achRate:
          type: number
        achRateYtd:
          type: number
        rankAch:
          type: integer
        statusWarna:
          type: string
        komponenDetail:
          type: ["string", "null"]
      required: [nik, namaAm, divisi, tahun, bulan, targetRevenue, realRevenue, achRate, achRateYtd, rankAch, statusWarna]
    PerformanceDetail:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        divisi:
          type: string
        summaries:
          type: array
          items:
            $ref: "#/components/schemas/PerformanceSummary"
      required: [nik, namaAm, divisi, summaries]
    FunnelLop:
      type: object
      properties:
        lopid:
          type: string
        judulProyek:
          type: string
        pelanggan:
          type: string
        nilaiProyek:
          type: number
        divisi:
          type: string
        statusF:
          type: string
        statusProyek:
          type: ["string", "null"]
        kategoriKontrak:
          type: ["string", "null"]
        estimateBulan:
          type: ["string", "null"]
        namaAm:
          type: string
        reportDate:
          type: string
      required: [lopid, judulProyek, pelanggan, nilaiProyek, divisi, statusF, namaAm, reportDate]
    FunnelOverview:
      type: object
      properties:
        totalLop:
          type: integer
        totalNilai:
          type: number
        targetFullHo:
          type: number
        realFullHo:
          type: number
        shortage:
          type: number
        cutoffDate:
          type: ["string", "null"]
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStatusGroup"
        byAm:
          type: array
          items:
            $ref: "#/components/schemas/FunnelAmGroup"
        lops:
          type: array
          items:
            $ref: "#/components/schemas/FunnelLop"
      required: [totalLop, totalNilai, targetFullHo, realFullHo, shortage, byStatus, byAm, lops]
    FunnelStatusGroup:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        totalNilai:
          type: number
      required: [status, count, totalNilai]
    FunnelAmGroup:
      type: object
      properties:
        namaAm:
          type: string
        nik:
          type: string
        divisi:
          type: string
        totalLop:
          type: integer
        totalNilai:
          type: number
        shortage:
          type: number
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStatusGroup"
      required: [namaAm, nik, divisi, totalLop, totalNilai, shortage, byStatus]
    FunnelDetail:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        divisi:
          type: string
        totalLop:
          type: integer
        totalNilai:
          type: number
        shortage:
          type: number
        lops:
          type: array
          items:
            $ref: "#/components/schemas/FunnelLop"
      required: [nik, namaAm, divisi, totalLop, totalNilai, shortage, lops]
...